fix(post_type check in admin): Fixed post_type check based on $post where $post was not available
Fixed undefined index warning because of not having actual $post object

+++ includes/ns-utility-functions.php
@@ -544,7 +544,7 @@ function ns_date_time( $datetime, $time = true ) {
      * @var  integer
      * ...
      */
-    if( ! (is_numeric($datetime) && (int) $datetime === $datetime) ) {
+    if( ! (is_numeric($datetime) && (float) $datetime === $datetime) ) {
         $datetime = strtotime($datetime);
     }
